Feat: Filtered by UserId(mock)

+++ src/domain/todo.go
@@ -1,7 +1,7 @@
 package domain
 
 type Todo struct {
-	Id int16 `json:"int"`
+	Id int `json:"int"`
 	UserId int16 `json:"int"`
 	Text string `json:"text"`
 }
